Factor out Kore.Step.Simplification.Predicate
