+++ kore/kore.cabal
@@ -444,6 +444,7 @@ library
     Kore.Step.Simplification.Overloading
     Kore.Step.Simplification.OverloadSimplifier
     Kore.Step.Simplification.Pattern
+    Kore.Step.Simplification.Predicate
     Kore.Step.Simplification.Rewrites
     Kore.Step.Simplification.Rule
     Kore.Step.Simplification.SetVariable
